11/18/2010:         Handle allocations of > 4GB -- reported by Howard Bussey         include/lib.h: dynamic(size_t)         lib/init.c: size_t ndynamic, pmstack, istack; getk takes size_t         lib/{bsd,generic,posix}/dynamic.c: dynamic(size_t)

+++ include/lib.h
@@ -124,7 +124,7 @@ void intspc __P((struct spec *, struct descr *));
 int spreal __P((struct descr *,struct spec *));
 
 /* from dynamic.c */
-char *dynamic __P((int));
+char *dynamic __P((size_t));
 void vm_gc_advise __P((int));
 
 /* from execute.c */
